abstract argentina tables from general models
